Switch to standard formula style
As recently discussed on salt-formulas it is preferred to keep settings (both
distribution specific ones and defaults for pillar data) in map.jinja and rely
on deep merging. This commit switches the formula to that style and thereby
provides a single way to access *all* settings via the collect_settings
dictionary.

Pillar backwards compatibility was mostly kept, but it was necessary to adjust
both the mysql and postgresql pillars to contain lists of dictionaries, rather
than simply defining a single db over which the previous code then "iterated".