Allow using local hostname for global Hostname
Note: this patch changes collectd's behaviour if grains['fqdn'] is not
the same as the output of hostname -f.  Restore the old behaviour by
setting this in the collectd piller:

  Hostname: {{ salt['grains.get']('fqdn') }}

With this patch Hostname is only set in collectd.conf if it is set in
the pillar data.  If it is not set collectd will determine it using
gethostname(2).  FQDNLookup is set in collectd.conf as follow:

- Hostname set: FQDNLookup not set (would be ignored by collectd)

- Hostname not set: FQDNLookup set from pillar data or default true